pyproject.toml
[tool.poetry]
name = "trjp"
version = "0.1.0"
description = "Source code repository for the AISTAT 2023 paper Transport Reversible Jump Proposals."
authors = ["Laurence Davies <laurence@latentlogic.com.au>"]
readme = "README.md"
packages = [
    { include = "./rjlab" },
    { include = "./examples" },
]

[tool.poetry.dependencies]
python = "^3.8"
nflows = "^0.14"
numba = "^0.56.4"
numpy = "^1.23.5"
torch = "^1.13.0"
seaborn = "^0.12.1"
setuptools = "^65.6.3"
matplotlib = "^3.6.2"
scipy = "^1.9.3"
scikit-learn = "^1.1.3"
pymc = "^4.4.0"
kdepy = "^1.1.0"
ott-jax = "^0.3.1"
dirichlet = {git = "https://github.com/ericsuh/dirichlet.git"}
black = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
